Backend/requirements.txt
Flask
flask_cors
mysql-connector-python
flask_jwt_extended
python-dotenv
requests 
openai
email_validator